sector/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
django==6.0.3
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ollama==0.6.1
pydantic==2.12.5
pydantic-core==2.41.5
pyjwt==2.12.1
python-decouple==3.8
sqlparse==0.5.5
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
